move FastMCP setup into app.py
