+++ run_server.py
@@ -1,15 +1,12 @@
 import asyncio
 import argparse
 import logging
-import os
 import yaml
-
-from fastapi import Request, Response
-from fastapi.responses import FileResponse, JSONResponse
-from mcp.server.fastmcp import FastMCP
 
 from rdkit_mcp.register_tools import register_tools
 from rdkit_mcp.settings import ToolSettings
+
+from rdkit_mcp.app import app
 
 logging.basicConfig(
     level=logging.INFO,
@@ -23,31 +20,6 @@ parser.add_argument("--transport", choices=["sse", "stdio"], help="Transport met
 parser.add_argument("--host", type=str, help="Host to run the server on", default="127.0.0.1")
 parser.add_argument("--settings", type=str, help="Path to YAML settings file", default=None)
 parser.add_argument("-v", "--verbose", action="store_true", help="Enable verbose (debug) logging")
-
-mcp = FastMCP("RDKit-MCP Server")
-
-
-@mcp.custom_route("/files", methods=["GET"])
-async def list_files(request: Request):
-    settings = ToolSettings()
-    file_dir = str(settings.FILE_DIR)
-    try:
-        files = os.listdir(file_dir)
-        return JSONResponse({"files": files})
-    except Exception as e:
-        return JSONResponse({"error": str(e)})
-
-
-@mcp.custom_route("/files/{filename}", methods=["GET"])
-async def get_file(request: Request):
-    filename = request.path_params.get("filename")
-    settings = ToolSettings()
-    file_dir = str(settings.FILE_DIR)
-    file_path = os.path.join(file_dir, filename)
-    if not os.path.isfile(file_path):
-        return JSONResponse(status_code=404, content={"error": "File not found"})
-    return FileResponse(path=file_path, filename=filename)    
-
 
 async def main():
     """Main function to run the MCP server."""
@@ -71,18 +43,18 @@ async def main():
     allow_list = settings.ALLOW_LIST
     block_list = settings.BLOCK_LIST
     logger.info("Registering tools with MCP server...")
-    await register_tools(mcp, allow_list=allow_list, block_list=block_list)
+    await register_tools(app, allow_list=allow_list, block_list=block_list)
 
     # Start server
     logger.info(f"Starting RDKit MCP Server with transport: {transport}")
     if transport == "sse":
         logger.info(f"Server running on {args.host}:{args.port} using SSE transport.")
-        mcp.settings.host = args.host
-        mcp.settings.port = args.port
-        await mcp.run_sse_async()
+        app.settings.host = args.host
+        app.settings.port = args.port
+        await app.run_sse_async()
     elif transport == "stdio":
         logger.info("Server running using stdio transport.")
-        await mcp.run_stdio_async()
+        await app.run_stdio_async()
 
 
 if __name__ == "__main__":
